rename: personnel-list -> personnel 변경
