+++ src/main/java/ministryofeducation/sideprojectspring/personnel/domain/department/Department.java
@@ -1,4 +1,4 @@
-package ministryofeducation.sideprojectspring.personnel_list.domain.department;
+package ministryofeducation.sideprojectspring.personnel.domain.department;
 
 public enum Department {
     KINDERGARTEN, HOLY_KIDS, PAUL_COMMUNITY, JOSHUA,
