+++ src/main/java/ministryofeducation/sideprojectspring/personnel/infrastructure/PersonnelListRepository.java
@@ -1,8 +1,7 @@
-package ministryofeducation.sideprojectspring.personnel_list.infrastructure;
+package ministryofeducation.sideprojectspring.personnel.infrastructure;
 
 import java.util.List;
-import java.util.Optional;
-import ministryofeducation.sideprojectspring.personnel_list.domain.Personnel;
+import ministryofeducation.sideprojectspring.personnel.domain.Personnel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PersonnelListRepository extends JpaRepository<Personnel, Long> {
