+++ src/main/java/ministryofeducation/sideprojectspring/personnel/presentation/PersonnelListController.java
@@ -1,9 +1,9 @@
-package ministryofeducation.sideprojectspring.personnel_list.presentation;
+package ministryofeducation.sideprojectspring.personnel.presentation;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
-import ministryofeducation.sideprojectspring.personnel_list.application.PersonnelListService;
-import ministryofeducation.sideprojectspring.personnel_list.presentation.dto.response.PersonnelListDto;
+import ministryofeducation.sideprojectspring.personnel.application.PersonnelListService;
+import ministryofeducation.sideprojectspring.personnel.presentation.dto.response.PersonnelListDto;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
